Add files via upload
Question 1: FizzBuzz
Write a program that prints the numbers from 1 to 100. For multiples of 3, print "Fizz"; for
multiples of 5, print "Buzz"; and for numbers that are multiples of both 3 and 5, print
"FizzBuzz".

Question 2: Fibonacci Sequence
Write a program to generate the Fibonacci sequence up to 100.
Question 3: Power of Two
Write a program that takes an integer as input and returns true if the input is a power of two.

Question 4: Capitalize Words
Write a program that accepts a string as input, capitalizes the first letter of each word in the
string, and then returns the result string.

question 5: Reverse Integer  Write a program that takes an integer as input and returns an integer with reversed digit ordering. 
 Question 6: Count Vowels Write a program that counts the number of vowels in a sentence.